requirements.txt
httpx==0.27.2
pydantic==2.9.2
SQLAlchemy==2.0.36
python-dotenv==1.0.1
APScheduler==3.10.4
python-telegram-bot[job-queue]==21.6
pytest==8.3.3
httpx-socks[asyncio]==0.9.1
geoip2==4.8.1
